Charging pins attached on bumper
If charging pins are part of bumper (like Tianchen mowers), bumpers will be triggered when robot reaches station. In most cases, mower will reverse and roll, so it leaves station.

By changing the state, the mower will always first check if charging pins are connected. In case of peri-track, if bumper gets triggered because mower reaches station, it will not reverse but change to correct state STATION